Fixed failing test cases
Broke test cases down by subject area instead of one large test file
