+++ .gitignore
@@ -14,3 +14,4 @@ gh-pages
 .vscode
 # Ignore Webstorm config files
 *.idea
+static/lib
